+++ test/unit/mocks/ipfs-mock.js
@@ -18,6 +18,8 @@ class MockIpfsInstance {
       }
     }
   }
+
+  stop () {}
 }
 
 module.exports = IPFS
